# Conflicts:
#	lib/main.dart
#	lib/track_my_indoor_exercise_app.dart
#	lib/ui/about.dart
#	lib/ui/activities.dart
#	lib/ui/calorie_tunes.dart
#	lib/ui/details/activity_details.dart
#	lib/ui/device_usages.dart
#	lib/ui/find_devices.dart
#	lib/ui/import_form.dart
#	lib/ui/leaderboards/device_leaderboard.dart
#	lib/ui/leaderboards/leaderboard_device_hub.dart
#	lib/ui/leaderboards/leaderboard_sport_hub.dart
#	lib/ui/leaderboards/leaderboard_type_picker.dart
#	lib/ui/leaderboards/sport_leaderboard.dart
#	lib/ui/parts/battery_status.dart
#	lib/ui/parts/boolean_question.dart
#	lib/ui/parts/calorie_factor_tune.dart
#	lib/ui/parts/calorie_override.dart
#	lib/ui/parts/color_picker.dart
#	lib/ui/parts/export_format_picker.dart
#	lib/ui/parts/heart_rate_monitor_pairing.dart
#	lib/ui/parts/heart_rate_monitor_scan_result.dart
#	lib/ui/parts/import_format_picker.dart
#	lib/ui/parts/legend_dialog.dart
#	lib/ui/parts/palette_picker.dart
#	lib/ui/parts/power_factor_tune.dart
#	lib/ui/parts/scan_result.dart
#	lib/ui/parts/spin_down.dart
#	lib/ui/parts/sport_picker.dart
#	lib/ui/parts/three_choices.dart
#	lib/ui/parts/upload_portal_picker.dart
#	lib/ui/power_tunes.dart
#	lib/ui/preferences/athlete.dart
#	lib/ui/preferences/data.dart
#	lib/ui/preferences/expert.dart
#	lib/ui/preferences/integrations.dart
#	lib/ui/preferences/preferences_hub.dart
#	lib/ui/preferences/row_configuration_dialog.dart
#	lib/ui/preferences/target_heart_rate.dart
#	lib/ui/preferences/user_experience_preferences.dart
#	lib/ui/preferences/workout.dart
#	lib/ui/preferences/zone_palette.dart
#	lib/ui/preferences/zones_hub.dart
#	lib/ui/recording.dart
#	lib/upload/constants.dart
#	lib/utils/preferences.dart
#	lib/utils/theme_manager.dart
#	pubspec.lock
#	pubspec.yaml

+++ test/crc_test.dart
@@ -195,7 +195,7 @@ void main() {
         0x03
       ], pad: false, crc: 0x9ed3),
     ]) {
-      final sum = testPair.data.fold<double>(0.0, (a, b) => a + b);
+      final sum = testPair.data.fold<int>(0, (a, b) => a + b);
       test("$sum -> ${testPair.crc}", () async {
         expect(crcData(testPair.data), testPair.crc);
       });
